+++ packages/moengage_personalize/moengage_personalize_android/android/src/main/kotlin/com/moengage/flutter/personalize/Constants.kt
@@ -1,0 +1,11 @@
+package com.moengage.flutter.personalize
+
+internal const val CHANNEL_NAME = "com.moengage/personalize"
+internal const val MODULE_TAG = "MoEFlutterPersonalize_"
+
+internal const val METHOD_FETCH_EXPERIENCES_META = "fetchExperiencesMeta"
+internal const val METHOD_FETCH_EXPERIENCES = "fetchExperiences"
+internal const val METHOD_EXPERIENCES_SHOWN = "experiencesShown"
+internal const val METHOD_EXPERIENCE_CLICKED = "experienceClicked"
+internal const val METHOD_OFFERINGS_SHOWN = "offeringsShown"
+internal const val METHOD_OFFERING_CLICKED = "offeringClicked"