+++ packages/moengage_personalize/moengage_personalize_platform_interface/lib/src/internal/payload_mapper.dart
@@ -43,6 +43,7 @@ Map<String, dynamic> campaignToMap(ExperienceCampaign campaign) {
     keyExperienceKey: campaign.experienceKey,
     keyPayload: campaign.payload,
     keyExperienceContext: campaign.experienceContext,
+    keySource: campaign.source.value
   };
 }